Refactor: Extract multi-dive planning and mutation logic to domain
Moves profile and cylinder editing logic directly into DivePlanInputModel in the domain layer, removing DiveEditorViewModelDelegate. Multi-dive calculation logic previously in PlanScreenViewModel is now extracted into MultiDivePlanner, also in the domain.
